@@ -1,9 +1,10 @@
 {
   "name": "openreview-web",
-  "version": "1.13.18",
+  "version": "1.14.1",
   "private": true,
   "scripts": {
     "dev": "next dev --port $NEXT_PORT",
+    "prod": "next build && next start --port $NEXT_PORT",
     "lint": "next lint",
     "build": "next build",
     "start": "next start --port $NEXT_PORT",
@@ -27,6 +28,7 @@
     "@codemirror/lint": "^6.4.2",
     "@codemirror/state": "^6.3.3",
     "@codemirror/view": "^6.22.2",
+    "@reduxjs/toolkit": "^2.5.0",
     "bootstrap": "^3.4.1",
     "codemirror": "^6.0.1",
     "copy-to-clipboard": "^3.3.3",
@@ -34,42 +36,46 @@
     "dayjs": "^1.11.10",
     "eslint-plugin-jest": "^27.6.0",
     "eslint4b-prebuilt": "^6.7.2",
+    "fluent-logger": "^3.4.1",
     "handlebars": "^4.7.8",
-    "isomorphic-dompurify": "^2.12.0",
+    "isomorphic-dompurify": "^2.21.0",
     "jquery": "^2.2.4",
     "jwt-decode": "^4.0.0",
     "lodash": "^4.17.21",
-    "marked": "^11.0.1",
+    "marked": "^15.0.6",
     "nanoid": "^5.0.4",
-    "next": "^13.5.6",
+    "next": "^15.2.3",
     "next-build-id": "^3.0.0",
     "query-string": "^7.1.3",
     "rc-picker": "^4.6.14",
     "rc-steps": "^6.0.1",
     "rc-virtual-list": "^3.11.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^19.0.0",
+    "react-error-boundary": "^4.1.2",
+    "react-redux": "^9.1.2",
     "react-select": "^5.8.0",
-    "sass": "^1.69.5",
+    "redux": "^5.0.1",
     "socket.io-client": "^4.8.1",
-    "swr": "^2.2.4",
-    "universal-cookie": "^6.1.1"
+    "universal-cookie": "^6.1.1",
+    "winston": "^3.17.0"
   },
   "devDependencies": {
     "@ffmpeg-installer/ffmpeg": "^1.1.0",
     "@next/env": "^13.5.6",
-    "@testing-library/jest-dom": "^6.1.5",
-    "@testing-library/react": "^14.1.2",
-    "@testing-library/user-event": "^14.5.1",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.2.0",
+    "@testing-library/user-event": "^14.6.1",
     "eslint": "^8.55.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-next": "^13.5.6",
+    "eslint-config-next": "^15.0.3",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.28.0",
+    "eslint-plugin-import": "^2.31.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "node-fetch-cjs": "^3.3.2",
     "prettier": "^3.1.0",
+    "sass": "^1.82.0",
     "testcafe": "^3.4.0"
   }
 }
