@@ -1,10 +1,6 @@
 /* globals promptError: false */
-import { useContext, useEffect, useState } from 'react'
-import { useRouter } from 'next/router'
-import useUser from '../../hooks/useUser'
-import useQuery from '../../hooks/useQuery'
-import { Tab, TabList, TabPanel, TabPanels, Tabs } from '../Tabs'
-import { referrerLink, venueHomepageLink } from '../../lib/banner-links'
+import { useContext, useEffect } from 'react'
+import { useSearchParams } from 'next/navigation'
 import WebFieldContext from '../WebFieldContext'
 import BasicHeader from './BasicHeader'
 import ErrorDisplay from '../ErrorDisplay'
@@ -12,6 +8,7 @@ import EthicsChairOverview from './EthicsChairConsole/EthicsChairOverview'
 import PaperStatus from './EthicsChairConsole/EthicsChairPaperStatus'
 import EthicsChairTasks from './EthicsChairConsole/EthicsChairTasks'
 import { getRoleHashFragment } from '../../lib/utils'
+import ConsoleTabs from './ConsoleTabs'
 
 const EthicsChairConsole = ({ appContext }) => {
   const {
@@ -28,38 +25,20 @@ const EthicsChairConsole = ({ appContext }) => {
     ethicsMetaReviewName,
     preferredEmailInvitationId,
   } = useContext(WebFieldContext)
-  const { setBannerContent } = appContext
-  const router = useRouter()
-  const query = useQuery()
-  const [activeTabId, setActiveTabId] = useState(
-    decodeURIComponent(window.location.hash) || '#overview'
-  )
-  const { user, userLoading } = useUser()
+  const { setBannerContent } = appContext ?? {}
+  const query = useSearchParams()
 
   const ethicsChairsUrlFormat = getRoleHashFragment(ethicsChairsName)
-  const validTabIds = [
-    '#overview',
-    `#${submissionName.toLowerCase()}-status`,
-    `#${ethicsChairsUrlFormat}-tasks`,
-  ]
 
   useEffect(() => {
     if (!query) return
 
-    if (query.referrer) {
-      setBannerContent(referrerLink(query.referrer))
+    if (query.get('referrer')) {
+      setBannerContent({ type: 'referrerLink', value: query.get('referrer') })
     } else {
-      setBannerContent(venueHomepageLink(venueId))
+      setBannerContent({ type: 'venueHomepageLink', value: venueId })
     }
   }, [query, venueId])
-
-  useEffect(() => {
-    if (!validTabIds.includes(activeTabId)) {
-      setActiveTabId(validTabIds[0])
-      return
-    }
-    router.replace(activeTabId)
-  }, [activeTabId])
 
   const missingConfig = Object.entries({
     header,
@@ -84,45 +63,29 @@ const EthicsChairConsole = ({ appContext }) => {
   return (
     <>
       <BasicHeader title={header?.title} instructions={header.instructions} />
-      <Tabs>
-        <TabList>
-          <Tab
-            id="overview"
-            active={activeTabId === '#overview' ? true : undefined}
-            onClick={() => setActiveTabId('#overview')}
-          >
-            Overview
-          </Tab>
-          <Tab
-            id={`${submissionName.toLowerCase()}-status`}
-            active={
-              activeTabId === `#${submissionName.toLowerCase()}-status` ? true : undefined
-            }
-            onClick={() => setActiveTabId(`#${submissionName.toLowerCase()}-status`)}
-          >
-            {submissionName} Status
-          </Tab>
-          <Tab
-            id={`${ethicsChairsUrlFormat}-tasks`}
-            active={activeTabId === `#${ethicsChairsUrlFormat}-tasks` ? true : undefined}
-            onClick={() => setActiveTabId(`#${ethicsChairsUrlFormat}-tasks`)}
-          >
-            Ethics Chair Tasks
-          </Tab>
-        </TabList>
-
-        <TabPanels>
-          <TabPanel id="overview">
-            <EthicsChairOverview />
-          </TabPanel>
-          <TabPanel id={`${submissionName.toLowerCase()}-status`}>
-            {activeTabId === `#${submissionName.toLowerCase()}-status` && <PaperStatus />}
-          </TabPanel>
-          <TabPanel id={`${ethicsChairsUrlFormat}-tasks`}>
-            {activeTabId === `#${ethicsChairsUrlFormat}-tasks` && <EthicsChairTasks />}
-          </TabPanel>
-        </TabPanels>
-      </Tabs>
+      <ConsoleTabs
+        defaultActiveTabId="overview"
+        tabs={[
+          {
+            id: 'overview',
+            label: 'Overview',
+            content: <EthicsChairOverview />,
+            visible: true,
+          },
+          {
+            id: `${submissionName.toLowerCase()}-status`,
+            label: `${submissionName} Status`,
+            content: <PaperStatus />,
+            visible: true,
+          },
+          {
+            id: `${ethicsChairsUrlFormat}-tasks`,
+            label: 'Ethics Chair Tasks',
+            content: <EthicsChairTasks />,
+            visible: true,
+          },
+        ]}
+      />
     </>
   )
 }
