@@ -9,16 +9,20 @@ let routerParams
 let noteSummaryProps
 let noteReviewStatusProps
 
-jest.mock('next/router', () => ({
+jest.mock('next/navigation', () => ({
+  useSearchParams: () => ({
+    get: () => jest.fn(),
+  }),
   useRouter: () => ({
-    replace: (params) => {
+    replace: jest.fn((params) => {
       routerParams = params
-      return jest.fn()
-    },
+      return {
+        catch: jest.fn(),
+      }
+    }),
   }),
 }))
 jest.mock('../hooks/useUser', () => () => useUserReturnValue)
-jest.mock('../hooks/useQuery', () => () => ({}))
 jest.mock('../components/webfield/NoteSummary', () => (props) => {
   noteSummaryProps(props)
   return <span>note summary</span>
@@ -29,6 +33,7 @@ jest.mock('../components/webfield/NoteReviewStatus', () => ({
     return <span>note review status</span>
   },
 }))
+jest.mock('../app/CustomBanner', () => () => <span>Custom Banner</span>)
 
 global.promptError = jest.fn()
 global.typesetMathJax = jest.fn()
@@ -36,6 +41,7 @@ global.DOMPurify = {
   sanitize: jest.fn(),
 }
 global.marked = jest.fn()
+global.$ = jest.fn(() => ({ on: jest.fn(), off: jest.fn(), modal: jest.fn() }))
 
 beforeEach(() => {
   useUserReturnValue = { user: {}, accessToken: 'some token' }
