@@ -661,7 +661,7 @@ export function pluralizeString(word) {
  * @param {string} roleName - role name to convert
  */
 export function getSingularRoleName(roleName) {
-  return roleName.endsWith('s') ? roleName.slice(0, -1) : roleName
+  return roleName?.endsWith('s') ? roleName.slice(0, -1) : roleName
 }
 
 /**
@@ -733,6 +733,22 @@ export function formatTasksData(
     return ''
   }
 
+  const noteHasAllRequiredFields = (invitation, note) => {
+    const requiredFields = Object.entries(invitation?.edit?.note?.content ?? {}).flatMap(
+      ([field, fieldDescription]) => {
+        if (
+          !fieldDescription.value?.param ||
+          fieldDescription.value.param.const ||
+          fieldDescription.value?.param?.optional === true
+        )
+          return []
+        return field
+      }
+    )
+
+    return requiredFields.every((field) => !isNil(note.content?.[field]?.value))
+  }
+
   // eslint-disable-next-line consistent-return
   const formatInvitation = (invitation) => {
     const inv = { ...invitation }
@@ -749,7 +765,15 @@ export function formatTasksData(
 
     if (inv.noteInvitation) {
       if (inv.details.repliedNotes?.length > 0) {
-        inv.completed = true
+        if (
+          inv.details.repliedNotes?.some((repliedNote) =>
+            noteHasAllRequiredFields(inv, repliedNote)
+          )
+        ) {
+          inv.completed = true
+        } else {
+          inv.hasMissingFields = true
+        }
       } else if (inv.details.repliedEdits?.length > 0) {
         const noteIds = new Set()
         const sortedEdits = inv.details.repliedEdits.sort((a, b) => a.tcdate - b.tcdate)
@@ -1179,7 +1203,7 @@ export function getDefaultTimezone() {
  * get meta invitation id to use when editing an invitation
  */
 export function getMetaInvitationId(invitation) {
-  const metaInvitationId = invitation?.details?.writableWith?.find((p) => p.edit === true)?.id
+  const metaInvitationId = `${invitation.domain}/-/Edit`
   return metaInvitationId
 }
 
@@ -1555,7 +1579,7 @@ export function getVenueTabCountMessage(unrepliedCommentCount, notDeployedCount)
 export function parseNumberField(value) {
   if (typeof value === 'string') {
     const valueString = value.substring(0, value.indexOf(':'))
-    return valueString ? parseInt(valueString, 10) : null
+    return valueString ? parseInt(valueString, 10) : parseInt(value, 10)
   } else if (typeof value === 'number') {
     return value
   }
@@ -1672,10 +1696,14 @@ export function getRejectionReasons(currentInstitutionName) {
       rejectionText: `The ORCID profile you've provided as a homepage is empty or does not match the Career & Education history you've provided.\n\n${instructionText}`,
     },
     {
-      value: 'lastNotice',
-      label: 'Last notice before block',
-      rejectionText: `If invalid info is submitted again, your email will be blocked.\n\n${instructionText}`,
+      value: 'requestVouch',
+      label: 'Request supervisor or coauthor to vouch',
+      rejectionText: `Please ask a supervisor or coauthor or colleague that already has an OpenReview profile to email us at info@openreview.net from the institutional email in their profile to vouch for your registration.\n\nPlease add the profile ID of the person vouching for you to the Relations section of your profile.\n\n${instructionText}`,
     },
   ]
   return rejectionReasons
+}
+
+export function formatGroupResults(apiRes) {
+  return (apiRes.groups?.[0]?.members || []).map((groupId) => ({ groupId, dueDate: null }))
 }
