@@ -1,9 +1,9 @@
 import { Selector } from 'testcafe'
 import { baseGroupId, conferenceGroupId } from './utils/api-helper'
 
-const activeVenues = Selector('#active-venues')
-const openVenues = Selector('#open-venues')
-const allVenues = Selector('#all-venues')
+const activeVenues = Selector('#active-venues').filterVisible()
+const openVenues = Selector('#open-venues').filterVisible()
+const allVenues = Selector('#all-venues').filterVisible()
 
 // eslint-disable-next-line no-unused-expressions
 fixture`Home page`.page`http://localhost:${process.env.NEXT_PORT}`
@@ -13,14 +13,14 @@ test('show active venues', async (t) => {
   await t
     // Active venues
     .expect(activeVenues.find('li').count)
-    .eql(4)
-    .expect(activeVenues.find('a').nth(0).textContent)
-    .eql(conferenceGroupName)
+    .eql(5)
     .expect(activeVenues.find('a').nth(1).textContent)
-    .eql(`Another${conferenceGroupName}`)
+    .eql(conferenceGroupName)
     .expect(activeVenues.find('a').nth(2).textContent)
-    .eql('ICLR 2021 Conference')
+    .eql(`Another${conferenceGroupName}`)
     .expect(activeVenues.find('a').nth(3).textContent)
+    .eql('ICLR 2021 Conference')
+    .expect(activeVenues.find('a').nth(4).textContent)
     .eql('TestVenue 2023 Conference')
 
     // Open for submissions
